Move AssociatedPerson to person.rs
